move python and javascript
